+++ android/src/main/kotlin/io/maido/intercom/IntercomFlutterPlugin.kt
@@ -34,7 +34,7 @@ class IntercomFlutterPlugin : FlutterPlugin, MethodCallHandler, EventChannel.Str
   }
 
   private val intercomPushClient = IntercomPushClient()
-  private lateinit var unreadConversationCountListener: UnreadConversationCountListener
+  private var unreadConversationCountListener: UnreadConversationCountListener? = null
 
   override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
     val channel = MethodChannel(flutterPluginBinding.getFlutterEngine().getDartExecutor(), "maido.io/intercom")
@@ -55,6 +55,13 @@ class IntercomFlutterPlugin : FlutterPlugin, MethodCallHandler, EventChannel.Str
         val appId = call.argument<String>("appId")
         Intercom.initialize(application, apiKey, appId)
         result.success("Intercom initialized")
+      }
+      call.method == "setBottomPadding" -> {
+        val padding = call.argument<Int>("bottomPadding")
+        if(padding != null) {
+          Intercom.client().setBottomPadding(padding)
+          result.success("Bottom padding set")
+        }
       }
       call.method == "setUserHash" -> {
         val userHash = call.argument<String>("userHash")
@@ -201,17 +208,73 @@ class IntercomFlutterPlugin : FlutterPlugin, MethodCallHandler, EventChannel.Str
     }
   }
 
+  // generate the user attributes
+  private fun getUserAttributes(call: MethodCall): UserAttributes {
+    // user attributes
+    val name = call.argument<String>("name")
+    val email = call.argument<String>("email")
+    val phone = call.argument<String>("phone")
+    val userId = call.argument<String>("userId")
+    val company = call.argument<String>("company")
+    val companyId = call.argument<String>("companyId")
+    val customAttributes = call.argument<Map<String, Any?>>("customAttributes")
+    val signedUpAt = call.argument<Any?>("signedUpAt")
+
+    val userAttributes = UserAttributes.Builder()
+
+    if (name != null) {
+      userAttributes.withName(name)
+    }
+
+    if (email != null) {
+      userAttributes.withEmail(email)
+    }
+
+    if (phone != null) {
+      userAttributes.withPhone(phone)
+    }
+
+    if (userId != null) {
+      userAttributes.withUserId(userId)
+    }
+
+    if (company != null && companyId != null) {
+      val icmCompany = Company.Builder()
+      icmCompany.withName(company)
+      icmCompany.withCompanyId(companyId)
+      userAttributes.withCompany(icmCompany.build())
+    }
+
+    if (customAttributes != null) {
+      for ((key, value) in customAttributes) {
+        userAttributes.withCustomAttribute(key, value)
+      }
+    }
+
+    val seconds: Long? = signedUpAt?.toString()?.toLongOrNull()
+    if (seconds != null)
+      userAttributes.withSignedUpAt(seconds)
+
+    return userAttributes.build()
+  }
+
   override fun onListen(arguments: Any?, events: EventChannel.EventSink?) {
     unreadConversationCountListener = UnreadConversationCountListener { count -> events?.success(count) }
-    Intercom.client().addUnreadConversationCountListener(unreadConversationCountListener)
+        .also {
+          Intercom.client().addUnreadConversationCountListener(it)
+        }
   }
 
   override fun onCancel(arguments: Any?) {
-    Intercom.client().removeUnreadConversationCountListener(unreadConversationCountListener)
+    if (unreadConversationCountListener != null) {
+      Intercom.client().removeUnreadConversationCountListener(unreadConversationCountListener)
+    }
   }
 
   override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
-    Intercom.client().removeUnreadConversationCountListener(unreadConversationCountListener)
+    if (unreadConversationCountListener != null) {
+      Intercom.client().removeUnreadConversationCountListener(unreadConversationCountListener)
+    }
   }
 
   override fun onDetachedFromActivity() {