+++ lib/intercom_flutter.dart
@@ -28,24 +28,29 @@ class Intercom {
     return _unreadChannel.receiveBroadcastStream();
   }
 
+  /// This method allows you to set a fixed bottom padding for in app messages and the launcher.
+  ///
+  /// It is useful if your app has a tab bar or similar UI at the bottom of your window.
+  /// [padding] is the size of the bottom padding in points.
+  static Future<dynamic> setBottomPadding(int padding) {
+    return _channel.invokeMethod('setBottomPadding', {'bottomPadding': padding});
+  }
+
   static Future<dynamic> setUserHash(String userHash) {
     return _channel.invokeMethod('setUserHash', {'userHash': userHash});
   }
 
-  static Future<dynamic> registerIdentifiedUser(
-      {String userId, String email}) {
+  static Future<dynamic> registerIdentifiedUser({String userId, String email}) {
     if (userId?.isNotEmpty ?? false) {
       if (email?.isNotEmpty ?? false) {
         throw ArgumentError(
             'The parameter `email` must be null if `userId` is provided.');
       }
-      return _channel
-          .invokeMethod('registerIdentifiedUserWithUserId', {
+      return _channel.invokeMethod('registerIdentifiedUserWithUserId', {
         'userId': userId,
       });
     } else if (email?.isNotEmpty ?? false) {
-      return _channel
-          .invokeMethod('registerIdentifiedUserWithEmail', {
+      return _channel.invokeMethod('registerIdentifiedUserWithEmail', {
         'email': email,
       });
     } else {
@@ -65,6 +70,7 @@ class Intercom {
     String company,
     String companyId,
     String userId,
+    int signedUpAt,
     String lang,
     Map<String, dynamic> customAttributes,
   }) {
@@ -75,6 +81,7 @@ class Intercom {
       'company': company,
       'companyId': companyId,
       'userId': userId,
+      'signedUpAt': signedUpAt,
       'lang': lang,
       'customAttributes': customAttributes,
     });
@@ -124,6 +131,7 @@ class Intercom {
   }
 
   static Future<dynamic> sendTokenToIntercom(String token) {
+    assert(token != null && token.isNotEmpty);
     print("Start sending token to Intercom");
     return _channel.invokeMethod('sendTokenToIntercom', {'token': token});
   }